+++ src/collective/contentalerts/tests/test_contentrules.py
@@ -7,6 +7,7 @@ from collective.contentalerts.interfaces import IAlert
 from collective.contentalerts.interfaces import IHasStopWords
 from collective.contentalerts.interfaces import IStopWords
 from collective.contentalerts.testing import COLLECTIVE_CONTENTALERTS_INTEGRATION_TESTING  # noqa
+from collective.contentalerts.testing import COLLECTIVE_CONTENTALERTS_FUNCTIONAL_TESTING  # noqa
 from plone import api
 from plone.app.contentrules.rule import Rule
 from plone.app.discussion.interfaces import IConversation
@@ -60,7 +61,8 @@ class TextAlertConditionTestCase(unittest.TestCase):
             container=self.portal,
             id='doc1',
             title='Document 1',
-            type='Document'
+            type='Document',
+            text='lala',
         )
 
     def _add_comment(self, text):
@@ -294,9 +296,9 @@ class TextAlertConditionTestCase(unittest.TestCase):
             container=self.portal,
             id='doc2',
             title='Document 2',
-            type='Document'
+            type='Document',
+            text='this gives one alert',
         )
-        document.text = 'this gives one alert'
         condition = TextAlertCondition()
         condition.stop_words = u'one alert\nanother alert'
 
@@ -350,9 +352,9 @@ class TextAlertConditionTestCase(unittest.TestCase):
             container=self.portal,
             id='doc2',
             title='Document 2',
-            type='Document'
+            type='Document',
+            text='this gives one alert'
         )
-        document.text = 'this gives one alert'
         condition = TextAlertCondition()
         condition.stop_words = u'one alert\nanother alert'
 
@@ -396,9 +398,9 @@ class TextAlertConditionTestCase(unittest.TestCase):
             container=self.portal,
             id='doc2',
             title='Document 2',
-            type='Document'
+            type='Document',
+            text='this gives one alert'
         )
-        document.text = 'this gives one alert'
         condition = TextAlertCondition()
         condition.stop_words = u'one alert\nanother alert'
 
@@ -479,9 +481,9 @@ class SpecificAlertConditionsTestCase(unittest.TestCase):
             container=self.portal,
             id='doc2',
             title='Document 2',
-            type='Document'
+            type='Document',
+            text='this gives one alert'
         )
-        document.text = 'this gives one alert'
         condition = InadequateTextAlertCondition()
         self._set_record_value(u'one')
 
@@ -496,9 +498,9 @@ class SpecificAlertConditionsTestCase(unittest.TestCase):
             container=self.portal,
             id='doc2',
             title='Document 2',
-            type='Document'
+            type='Document',
+            text='this gives one alert'
         )
-        document.text = 'this gives one alert'
         condition = ForbiddenTextAlertCondition()
         self._set_record_value(u'one', record='forbidden_words')
 
@@ -510,7 +512,7 @@ class SpecificAlertConditionsTestCase(unittest.TestCase):
 
 
 class ContentRulesSubstitutionsTest(unittest.TestCase):
-    layer = COLLECTIVE_CONTENTALERTS_INTEGRATION_TESTING
+    layer = COLLECTIVE_CONTENTALERTS_FUNCTIONAL_TESTING
 
     def setUp(self):
         setupCoreSessions(self.layer['app'])
@@ -523,7 +525,8 @@ class ContentRulesSubstitutionsTest(unittest.TestCase):
             container=self.portal,
             id='doc1',
             title='Document 1',
-            type='Document'
+            type='Document',
+            text='lala'
         )
 
     def _add_comment(self, text='lilala'):